+++ backend/src/test/java/com/example/backend/domain/admin/AdminControllerTest.kt
@@ -9,10 +9,7 @@ import com.example.backend.domain.member.entity.Member
 import com.example.backend.domain.member.repository.MemberRepository
 import jakarta.servlet.http.Cookie
 import org.hibernate.validator.internal.util.Contracts.assertNotNull
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
+import org.junit.jupiter.api.*
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.MediaType
@@ -64,6 +61,7 @@ class AdminControllerTest(
 
         @BeforeAll
         fun setUp() {
+            adminRepository.deleteAll()
             val admin = Admin("admin", "\$2a\$12\$wS8w9vGzZ345XlGazbp8mekCkPyKoPFbky96pr0EqW.6I0Xtdt.YO")
             adminRepository.save(admin)
         }
@@ -134,5 +132,10 @@ class AdminControllerTest(
 
         val group = groupRepository.findById(groupId).orElseThrow()
         assert(group.status == GroupStatus.DELETED)
+    }
+
+    @AfterAll
+    fun finish() {
+        adminRepository.deleteAll()
     }
 }
